Change requests for Rust update v1.66.0
error: casting to the same type is unnecessary

Test: ./build.py --lto=thin
Bug: 263153841

+++ host/commands/secure_env/rust/ffi.rs
@@ -29,7 +29,7 @@ pub extern "C" fn kmr_ta_main(
     security_level: c_int,
     trm: *mut libc::c_void,
 ) {
-    let security_level = match security_level as i32 {
+    let security_level = match security_level {
         x if x == SecurityLevel::TrustedEnvironment as i32 => SecurityLevel::TrustedEnvironment,
         x if x == SecurityLevel::Strongbox as i32 => SecurityLevel::Strongbox,
         x if x == SecurityLevel::Software as i32 => SecurityLevel::Software,
